feat: support wrapper types as well
